move all packages into internal folder
